Implement versioned shader accessor (#3)
* Implement versioned shader accessor

Creates a new "ShaderAccessor" interface and implements it across various versions. It also adds new Optifine mappings for different versions and updates the "OptiFineAddon" to initialize the screen handler from the reference storage. It allows for version-specific shader access, improving mod compatibility with different versions of Minecraft.

* Use production release

* Update labyGradlePluginVersion and refactor OptiFineScreen

Updated the labyGradlePluginVersion from 0.3.40 to 0.3.41 in the settings file. Refactored the OptiFineScreen class by replacing the boolean 'options' attribute with a 'ScreenTag' attribute, allowing for more accurate tagging of screen options.